requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.0
django-crontab==0.7.1
django-redis==5.3.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
